requirements.txt
pandas
matplotlib
numpy<2
scikit-learn
openai
tiktoken
loguru
ipykernel
tqdm
pytesseract
pillow
evaluate
jiwer
requests
tabulate
python-dotenv
boto3
transformers
diskcache
google-cloud-vision
omegaconf
google-genai